better auto-coloring behaviour
This is a change that breaks backwards compatibility. However,
it only does so because the previous behavior is considered to
be a bug, not a feature. Now, when colored => 'auto' (the default
behavior) Data::Printer will respect the ANSI_COLORS_DISABLED
environment variable. In colored => 'auto' mode, it will also
colorize only when it's not returning, but printing to STDERR.
In 'auto', saying "my $var = p($something)" will remove colors.

As a consequence, this should help address the issues in MS Windows.