control the return value of the p() function (mst++)
There are times when a user might want more control over the
return value of p(). This can be set via the new 'return_value'
property. Available options are 'void', 'dump', and 'pass'
(for pass-through).

Default is 'dump', which behaves exactly like previous versions.
Note that, as usual, the 'dump' mode will only print the dump if
called in void context. Otherwise, it will only return it.