we can't guarantee sort other in bleadperl
it's actually something of a tiny miracle that this test never failed on
cpantesters. Now, with bleadperl's hash randomization, it has a mich higher
likelyhood of blowing up. Since there's a chance the randomization will
pick the keys in a sorted fashion, we test the other way around, so now
the test will blow up only if keys aren't sorted. This, of course, is
very silly since we actually rely on sorting for most of our other
print tests. So, until we refactor to enable proper testing of unsorted
methods, we're gonna leave just the sorted test.