MAJOR REFACTORING! Underlying object now available
Data::Printer now uses Data::Printer::Object, which is exposed in
filters and can be used elsewhere too. It has a simple public interface
(which shall be thoroughly documented in the following commits),
supports themes, refcounting, memory size and a bunch of neat features.
The only backwards compatibility that we have purposedly broken is the
.dataprinter file, which is now a simpler key=value pair text file. As a
consequence, you will not be able to put custom filters in there: you
have to create a standalone filter and load that via the .dataprinter.