accept leading/trailing whitespace in numbers under lax mode
when a string contains just whitespace and a number, it will
be interpreted by perl as exactly that number. For instance,
"123   " + 1 is 124, and "   321" + 1 is 322. So we must allow
it undex lax mode as a number. In strict mode, it will still show
as a dualvar containing both the original string and its numification.