Remove JSON::Parser as a dependency.
JSON::Parser is no longer available on CPAN. It was removed from the
JSON module in 2010. Including it as a dependency can cause the travis
test to install the JSON module version 1.15 from 2007 (which includes
JSON::Parser as a sub module). Incidentally this also installs an old
version of JSON::PP (also included in the 2007 JSON module) which can
prevent the installation of a newer version of JSON::PP which again
can cause test failures (see e.g. test t/104-filter_web.t).