Remove seen entries when their object is destroyed
The code was previously assuming refaddrs would never be reused, however
this not true when dealing with tied hashes (because the values they
return are temporaries). This avoids this problem by using a piece of
fieldhash infrastructure to delete entries from the seen hash when the
associated object is destroyed.

Using a fieldhash would have been the easiest solution but due to a bug
in how its implementation deals with objects that have string
overloading we can't use that here.